(UGLY patch ahead) Fixed UCX problem
"ucp_stream_worker_poll" will not signal two times the same endpoint if no progress on that endpoint is performed.
Hence, we need to be sure that when yielding the handle, no more data are present on the endpoint.
In case more data are available (we probe before yielding), the handle MUST go back in the Manager ready queue.